requirements.txt
asgiref==3.7.2
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.15.0
djangorestframework==3.14.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
pytest-django==4.5.2
pytz==2023.3.post1
sqlparse==0.4.4
